Add files via upload
This is a project that was completed while taking the Udemy course, Python for Machine Learning & Data Science Masterclass. The GOAL: To create a classification model that can predict whether or not a person has the presence of heart disease based on the physical features of that person (age,sex, cholesterol, etc.).